chore: refactor router
